Bug 1290789: firefox_primary_builds.json should have release and beta values for en-US
firefox_primary_builds.json does not have beta and release fields for en-US because
those are built with generateLocalizedBuilds() and based on the existence of an
l10n changeset.

The fix is to add an artificial changeset for en-US in generateLocalizedBuilds() because we will always ship en-US.

The tests added for en-US cover all channels. Note that there is a separate bug about the esr suffix missing, that is Bug 1290725.
That means that I am not including the esr suffix in the test, I will fix the test with Bug 1290725, ex:
self.assertTrue('2.0.2' in primary['en-US'])
will become once Bug 1290725 is fixed:
self.assertTrue('2.0.2esr' in primary['en-US'])