Improve tests on prerelease version numbers in config.py
Currently our tests compare the version numbers with version numbers defined in the tests themselves.
This is not useful, we should test what can fail when we modify the values in config.py, that is not the version numbers but the allowed patterns for Nightly and Aurora for a version number to avoid human errors such as a trailing space.